调整英雄信息面板，支持英雄属性信息显示

+++ uc/Assets/TangGame/Scripts/UI/Panel/HeroViewPanel.cs
@@ -83,7 +83,7 @@ namespace TangGame.UI
 		void ShowItemByHeroLocation (HeroLocationEnum em)
 		{
 			foreach (HeroItem item in heroItems.Values) {
-				if (item.Data.Xml.location.Equals (em)) {
+				if (item.Data.Xml.location.Equals ((int)em)) {
 					item.gameObject.SetActive (true);
 				}
 			}
